fixed one way flight bugs

+++ src/Twig/Components/DeleteFlight.php
@@ -20,28 +20,27 @@ class DeleteFlight extends AbstractController
     #[LiveProp]
     public int $id;
 
-    #[LiveProp]
-    public int $id_return;
-
     #[LiveAction]
-    public function deleteFlight(#[LiveArg] int $id, #[LiveArg] int $idReturn, EntityManagerInterface $entityManager)
+    public function deleteFlight(#[LiveArg] int $id, #[LiveArg] EntityManagerInterface $entityManager)
     {
 
         $flight = $entityManager->getRepository(Flight::class)->find($id);
-        $flightReturn = $entityManager->getRepository(Flight::class)->find($idReturn);
         $prices = $entityManager->getRepository(FlightPrices::class)->findby(['flight_id' => $id]);
-        $pricesReturn = $entityManager->getRepository(FlightPrices::class)->findby(['flight_id' => $idReturn]);
-
         $entityManager->remove($flight);
-        $entityManager->remove($flightReturn);
         foreach ($prices as $price) {
             $entityManager->remove($price);
         }
-        foreach ($pricesReturn as $priceReturn) {
-            $entityManager->remove($priceReturn);
+
+        if (isset($flight->return_flight)) {
+            $flightReturn = $entityManager->getRepository(Flight::class)->find($flight->return_flight);
+            $pricesReturn = $entityManager->getRepository(FlightPrices::class)->findby(['flight_id' => $flight->return_flight]);
+            $entityManager->remove($flightReturn);
+            foreach ($pricesReturn as $priceReturn) {
+                $entityManager->remove($priceReturn);
+            }
         }
 
-        $entityManager->flush();
+       $entityManager->flush();
 
         return $this->redirectToRoute('app_flights');
